Done line avoidance code

+++ lib/Bluetooth/Bluetooth.cpp
@@ -4,9 +4,10 @@ void Bluetooth::init(){
 	blueSerial.begin(38400);
 }
 
-void Bluetooth::send(double cmd){
+void Bluetooth::send(uint8_t cmd){
 	blueSerial.write(cmd);
 }
+
 double Bluetooth::receive(){
 	if (blueSerial.available())
 	{
